refactor: typescript
