--- index.d.ts
@@ -1,1 +1,0 @@
-export default function remarkDetails(): void;
