--- micromark-extension-details/html.d.ts
@@ -1,6 +1,0 @@
-export function detailsHtml(): Record<"enter" | "exit", {
-	detailsContainer(): void,
-	detailsContainerClassName(): void,
-	detailsContainerSummary(): void,
-	detailsContainerContent(): void
-}>
