--- micromark-extension-details/index.js
@@ -1,3 +1,0 @@
-export { detailsHtml } from './html.js';
-import syntax from './syntax.js';
-export default syntax;
