+++ micromark-extension-details/index.ts
@@ -1,0 +1,2 @@
+export { detailsHtml } from './html.js';
+export { default as syntax } from './syntax.js';
