--- micromark-extension-details/syntax.d.ts
@@ -1,6 +1,0 @@
-export default function syntax(): {
-	flow: Record<number, {
-		tokenize(effects: any, ok: any, nok: any): (code: any) => any,
-		concrete: boolean,
-	}>
-};
